requirements.txt
Flask==2.0.3
numpy==1.21.2
Pillow==8.4.0
tensorflow==2.6.0
keras==2.6.0
matplotlib==3.4.3